pyproject.toml
[tool.poetry]
name = "ava-mosaic-ai"
version = "0.0.0"
description = ""
authors = ["Karan Singh Kochar <pyrotank41@gmail.com>"]
readme = "README.md"
homepage = "https://github.com/pyrotank41/Mosaic-AI.git"
repository = "https://github.com/pyrotank41/Mosaic-AI.git"
packages = [{include = "ava_mosaic_ai"}]

[tool.poetry.dependencies]
python = "^3.11"
instructor = "^1.3.7"
pydantic = "^2.8.2"
anthropic = "^0.34.0"
openai = "^1.41.0"
python-dotenv = "^1.0.1"
pytest = "^8.3.2"


[tool.poetry.group.dev.dependencies]
ipykernel = "^6.29.5"
poetry-dynamic-versioning = "^1.4.0"
vcrpy = "^6.0.1"

[tool.poetry-dynamic-versioning]
enable = true
vcs = "git"
style = "semver"

[build-system]
requires = ["poetry-core>=1.0.0", "poetry-dynamic-versioning>=1.0.0,<2.0.0"]
build-backend = "poetry_dynamic_versioning.backend"
